Finalized Restructuring of Pipeline object
