+++ final_model.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pandas as pd
 from sklearn.decomposition import TruncatedSVD
-from sklearn.metrics import mean_squared_error
 from sklearn.preprocessing import LabelEncoder
+from sklearn.model_selection import train_test_split
 from scipy.sparse import csr_matrix
 
 user_data = pd.read_csv("data/Users.csv")
@@ -18,59 +18,67 @@ user_rating_books_ds = user_rating_books_ds[Feature_List]
 
 average_ratings = pd.read_csv("./data/average_ratings.csv.gz", delimiter = '\t', index_col = 0)
 
-
 class RecommenderPipeline:
     #Encodes data and creates user-item matrix
     def __init__(self):
-        self.user_rating = user_rating_books_ds
-        self.model = TruncatedSVD(n_components=50)
-        self.train_data = None
-        self.test_data = None
-        # Encode users and books
-        user_encoder = LabelEncoder()
-        book_encoder = LabelEncoder()
+        self.model = TruncatedSVD(n_components=2)
+        self.predicted_matrix = None
+        self.iteraction_matrix = None
+        self.book_encoder = None
         
-        self.user_rating.loc[:, "User-ID"] = user_encoder.fit_transform(self.user_rating["User-ID"]).astype(np.int32)
-        self.user_rating.loc[:, "Book-Title"] = book_encoder.fit_transform(self.user_rating["Book-Title"]).astype(np.int32)
-        self.user_rating.loc[:, "Book-Rating"] = self.user_rating["Book-Rating"].astype(np.int8)
+        # Encode users and books
+        #A sample of taken to prevent intense runtimes
+        user_rating = user_rating_books_ds.sample(frac=0.05)
+        user_encoder = LabelEncoder()
+        self.book_encoder = LabelEncoder()
+        
+        user_rating.loc[:, "User-ID"] = user_encoder.fit_transform(user_rating["User-ID"]).astype(np.int32)
+        user_rating.loc[:, "Book-Title"] = self.book_encoder.fit_transform(user_rating["Book-Title"]).astype(np.int32)
+        user_rating.loc[:, "Book-Rating"] = user_rating["Book-Rating"].astype(np.int8)
 
-        # Create user-item interaction matrix
+        # Create interaction matrix
         self.interaction_matrix = csr_matrix(
-            (user_rating_books_ds["Book-Rating"],
-            (user_rating_books_ds["User-ID"], user_rating_books_ds["Book-Title"]))
+            (user_rating["Book-Rating"],
+            (user_rating["User-ID"], user_rating["Book-Title"]))
         )
         
     #Test/Train splits data and trains model using train data
     def fit(self):
-        # Train-test split
-        train_ratio = 0.8
-        train_size = int(self.interaction_matrix.shape[0] * train_ratio)
-        self.train_data = self.interaction_matrix[:train_size].astype(np.float32)
-        self.test_data = self.interaction_matrix[train_size:].astype(np.float32)
+        non_zero_indices = np.transpose(self.interaction_matrix.nonzero()) 
+        ratings = self.interaction_matrix[non_zero_indices[:, 0], non_zero_indices[:, 1]].A1  
+        train_indices, test_indices = train_test_split(range(len(ratings)), test_size=0.2, random_state=42)
         
-        self.model.fit(self.train_data)
+        train_matrix = self.interaction_matrix.copy()
 
-    # This function creates k recommendations for a given user 
-    def recommend(self,user_id,num_recommendations):
-        user_ratings = self.trainset.ur[self.trainset.to_inner_uid(user_id)]
-        all_items = set(self.trainset.all_items())
-        rated_items = set(item for item, _ in user_ratings)
-        unrated_items = all_items - rated_items
+        for index in test_indices:
+            train_matrix[non_zero_indices[index][0], non_zero_indices[index][1]] = 0
+            
+        user_factors = self.model.fit_transform(train_matrix)
+        item_factors = self.model.components_
+
+        self.predicted_matrix = np.dot(user_factors, item_factors)
+
+    #This function creates num_recommendations recommendations for a given user 
+    def recommend(self, user_index, num_recommendations = 20):
         
-        recommendations = [(self.trainset.to_raw_iid(item), self.model.predict(user_id, self.trainset.to_raw_iid(item)).est)
-            for item in unrated_items]
+        user_predictions = self.predicted_matrix[user_index, :]
     
-        recommendations.sort(key=lambda x: x[1], reverse=True)
-        recommendations = recommendations[:num_recommendations]   
+        #Recursively find and select the indices of the top N items (highest predicted ratings)
+        top_n_items = np.argsort(user_predictions)[::-1][:num_recommendations]
+        decoded_categories = self.book_encoder.inverse_transform(top_n_items)
+        top_ratings = user_predictions[top_n_items]
+        recommendations = list(zip(decoded_categories, top_ratings))
+        
 
         #Adds ISBN and returns recommendation dataframe 
         recommendations = pd.DataFrame(recommendations)
         recommendations = pd.concat([recommendations, pd.DataFrame(books_ID_list['ISBN'])], join = 'inner', axis = 1)
-        recommendations = recommendations.rename(columns = {0 : "Book-Title", 1: "Rating"})
+        recommendations = recommendations.rename(columns = {0 : "Book-Title"})
+        recommendations = recommendations.drop(columns = {1})
         return recommendations
 
     #For users with no previous data, recommend a random list of books by taking a proportion of books within each rating range
-    def cold_recommend(self, num_recommendations): 
+    def cold_recommend(self, num_recommendations = 20): 
         fours = average_ratings.loc[average_ratings['Rating'] >= 4.0]   # 5%
         fours = fours.loc[fours['Rating'] < 5.0]
         
@@ -108,9 +116,8 @@ class RecommenderPipeline:
         final_df = pd.concat([fours, fives, sixes, sevens, eights, nines], axis = 0)
         final_df = final_df.sample(frac = 1)
         final_df = final_df.reset_index(drop=True)
+        final_df = final_df.drop(columns = {'Rating'})
         return final_df
+    
 
-
-#user_id = 10
-#num_recommendations = 20
-#print(f"Recommendations for User {user_id}: {recommend(user_id, num_recommendations)}")
+#15996 is the largest valid user_index
